requirements.txt
flask==3.1.1
python-dotenv==1.0.0
requests==2.32.5
instaloader>=4.10
pyobjc-framework-Vision
gspread>=6.0
